prefinal-1

+++ newsletters/views.py
@@ -7,66 +7,21 @@ from .models import NewsletterSubscription
 
 # Create your views here.
 @login_required
-@login_required
 def subscribe(request, author_id):
-
-    author = get_object_or_404(
-        User,
-        pk=author_id
-    )
-
-    NewsletterSubscription.objects.get_or_create(
-        subscriber=request.user,
-        author=author
-    )
-
-    messages.success(
-        request,
-        "Subscribed successfully."
-    )
-
-    return redirect(
-        "accounts:public_profile",
-        username=author.username
-    )
+    author = get_object_or_404(User,pk=author_id)
+    NewsletterSubscription.objects.get_or_create(subscriber=request.user,author=author)
+    messages.success(request,"Subscribed successfully.")
+    return redirect("accounts:public_profile",username=author.username)
 
 @login_required
 def unsubscribe(request, author_id):
-
-    author = get_object_or_404(
-        User,
-        pk=author_id
-    )
-
-    NewsletterSubscription.objects.filter(
-        subscriber=request.user,
-        author=author
-    ).delete()
-
-    messages.success(
-        request,
-        "Unsubscribed successfully."
-    )
-
-    return redirect(
-        "accounts:public_profile",
-        username=author.username
-    )
+    author = get_object_or_404(User,pk=author_id)
+    NewsletterSubscription.objects.filter(subscriber=request.user,author=author).delete()
+    messages.success(request,"Unsubscribed successfully.")
+    return redirect("accounts:public_profile",username=author.username)
 
 
 @login_required
 def my_subscriptions(request):
-
-    subscriptions = (
-        NewsletterSubscription.objects
-        .filter(subscriber=request.user)
-        .select_related("author")
-    )
-
-    return render(
-        request,
-        "newsletters/my_subscriptions.html",
-        {
-            "subscriptions": subscriptions
-        }
-    )
+    subscriptions = (NewsletterSubscription.objects.filter(subscriber=request.user).select_related("author"))
+    return render(request,"newsletters/my_subscriptions.html",{"subscriptions": subscriptions})
